+++ libs/pavex/src/server/server.rs
@@ -6,9 +6,47 @@ use crate::server::server_handle::ServerHandle;
 
 use super::IncomingStream;
 
-/// A builder for [`ServerHandle`]s.
+/// An HTTP server to handle incoming connections for Pavex applications.
 ///
-/// Check out [`ServerHandle::builder`] for more information.
+/// # Example
+///
+/// ```rust
+/// use std::net::SocketAddr;
+/// use pavex::server::Server;
+///
+/// # struct ApplicationState;
+/// # async fn router(_req: hyper::Request<hyper::body::Incoming>, _state: ApplicationState) -> pavex::response::Response { todo!() }
+/// # async fn t() -> std::io::Result<()> {
+/// # let application_state = ApplicationState;
+/// let addr = SocketAddr::from(([127, 0, 0, 1], 8080));
+///
+/// Server::new()
+///     .bind(addr)
+///     .await?
+///     // Both the routing function and the application state will usually
+///     // be code-generated by Pavex starting from your Blueprint.
+///     // You don't have to define them manually!
+///     .serve(router, application_state)
+///     # ;
+///     // [...]
+/// # Ok(())
+/// # }
+/// ```
+///
+/// # Configuration
+///
+/// You can customize the server default settings by creating your own [`ServerConfiguration`]
+/// and invoking [`Server::set_config`].
+///
+/// # Architecture
+///
+/// By default, [`Server::serve`] creates a worker per CPU core and distributes connection from an
+/// acceptor thread using a round-robin strategy.
+///
+/// Each worker has its own single-threaded [`tokio`] runtime—there is no work stealing across
+/// workers.  
+/// Each worker takes care to invoke your routing and request handling logic, with the help
+/// of [`hyper`].
 #[must_use = "You must call `serve` on a `Server` to start listening for incoming connections"]
 pub struct Server {
     config: ServerConfiguration,
